feat: restructure pages
