+++ web/src/components/Variants/ProteinCard.tsx
