--- web/src/pages/about.tsx
@@ -1,1 +1,0 @@
-export { AboutPage as default } from 'src/components/About/AboutPage'
