+++ web/src/pages/index.tsx
@@ -1,1 +1,1 @@
-export { MainPage as default } from 'src/components/Main/MainPage'
+export { HomePage as default } from 'src/components/Home/HomePage'
