+++ web/src/pages/variants/index.tsx
@@ -1,0 +1,1 @@
+export { VariantsPage as default } from 'src/components/Variants/VariantsPage'
